% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractCommercialPaperSeries}
\alias{extractCommercialPaperSeries}
\title{Extract commercial paper series}
\usage{
extractCommercialPaperSeries(x, seriesIndex, columnName, progressBar = "none")
}
\arguments{
\item{x}{commercial paper data sets as from \code{\link{getCommercialPaper}}.}

\item{seriesIndex}{the series number from within the data set}

\item{columnName}{name for the extracted observation value column}

\item{progressBar}{optional progress bar for \code{llply} progress monitoring, default none; see \code{plyr::create_progress_bar}}
}
\value{
list containing prepared date and series data frame elements
}
\description{
Extract commercial paper series
}
\details{
Extracts a given series from the commercial paper data sets.
Not meant to be used directly, but useful if class name assignment unwanted.
Substitutes observation value \code{NA} for any off-nominal observation status indicators.
}
\seealso{
\code{\link{getCommercialPaper}},
\code{\link{mergeSeries}}
}

