% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCommercialPaper}
\alias{getCommercialPaper}
\title{Get commercial paper history.}
\usage{
getCommercialPaper()
}
\value{
Returns the many data series as a list of class \code{usfrbcp}.
}
\description{
Fetches the US Federal Reserve Bank's commercial paper data, parses and returns as list.
}
\details{
Downloads from the Federal Reserve Bank's web service a zip file of several XML and XSD files
  related to commercial paper rates, volumes, and outstanding quantities over time.  Extracts the
  commercial paper XML files in a temporary directory and parses its XML into a list. The web service
  URL used is \url{http://www.federalreserve.gov/datadownload/Output.aspx?rel=CP&filetype=zip}.
}
\examples{
\donttest{
cp <- getCommercialPaper()
summary(cp)
}
}
\author{
Matt Barry <mrb@softisms.com>
}
\seealso{
\code{\link{getCommercialPaperRates}} for extracting commercial paper rates,
\code{\link{getCommercialPaperRatesOld}} for extracting old commercial paper rates,
\code{\link{getCommercialPaperVolumes}} for extracting commercial paper volumes,
\code{\link{getCommercialPaperOutstanding}} for extracting commercial paper outstanding,
\code{\link{getCommercialPaperOutstandingOld}} for extracting old commercial paper outstanding,
\code{\link{getCommercialPaperYearend}} for extracting commercial paper outstanding at year end,
\code{\link{mergeSeries}} for merging list elements into a data frame.
}

