% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCommercialPaperOutstanding}
\alias{getCommercialPaperOutstanding}
\title{Extract the commercial paper outstanding data.}
\usage{
getCommercialPaperOutstanding(cp, progressBar = "none")
}
\arguments{
\item{cp}{commercial paper data as retrieved by \code{getCommercialPaper()}.}

\item{progressBar}{optional progress bar for \code{llply} progress monitoring,
default none; see {\code{plyr::create_progress_bar}}.}
}
\value{
list containing data frame for each series, class \code{usfrbcpo}.
}
\description{
Extracts the commercial paper outstanding series from the commercial paper data set.
}
\examples{
\donttest{
cp <- getCommercialPaper()
outstanding <- getCommercialPaperOutstanding(cp,"text")
summary(outstanding)
}
}
\seealso{
\code{\link{getCommercialPaper}},
\code{\link{getCommercialPaperRates}},
\code{\link{getCommercialPaperRatesOld}},
\code{\link{getCommercialPaperVolumes}},
\code{\link{getCommercialPaperOutstandingOld}},
\code{\link{getCommercialPaperYearend}},
\code{\link{mergeSeries}}
}

