% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCommercialPaperRates}
\alias{getCommercialPaperRates}
\title{Extract the commercial paper rates data set}
\usage{
getCommercialPaperRates(cp, progressBar = "none")
}
\arguments{
\item{cp}{commercial paper data as retrieved by \code{getCommercialPaper()}}

\item{progressBar}{optional progress bar for \code{llply} progress monitoring,
default none; see {\code{plyr::create_progress_bar}}}
}
\value{
list containing data frame for each series, class \code{usfrbcpr}
}
\description{
Extract the commercial paper rates data set
}
\examples{
\dontrun{
cp <- getCommercialPaper()
rates <- getCommercialPaperRates(cp,"text")
summary(rates)
}
}
\seealso{
\code{\link{getCommercialPaper}},
\code{\link{getCommercialPaperRatesOld}},
\code{\link{getCommercialPaperVolumes}},
\code{\link{getCommercialPaperOutstanding}},
\code{\link{getCommercialPaperOutstandingOld}},
\code{\link{getCommercialPaperYearend}},
\code{\link{mergeSeries}}
}

