% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCommercialPaperVolumes}
\alias{getCommercialPaperVolumes}
\title{Extract the commercial paper volumes data set}
\usage{
getCommercialPaperVolumes(cp, progressBar = "none")
}
\arguments{
\item{cp}{commercial paper data as retrieved by \code{getCommercialPaper()}}

\item{progressBar}{optional progress bar for \code{llply} progress monitoring,
default none; see {\code{plyr::create_progress_bar}}}
}
\value{
list containing data frame for each series, class \code{usfrbcpv}
}
\description{
Extract the commercial paper volumes data set
}
\examples{
\dontrun{
cp <- getCommercialPaper()
volumes <- getCommercialPaperVolumes(cp,"text")
summary(volumes)
}
}
\seealso{
\code{\link{getCommercialPaper}},
\code{\link{getCommercialPaperRates}},
\code{\link{getCommercialPaperRatesOld}},
\code{\link{getCommercialPaperYearend}},
\code{\link{getCommercialPaperOutstanding}},
\code{\link{getCommercialPaperOutstandingOld}},
\code{\link{mergeSeries}}
}

