% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCommercialPaperYearend}
\alias{getCommercialPaperYearend}
\title{Extract the commercial paper outstanding year end data set}
\usage{
getCommercialPaperYearend(cp, progressBar = "none")
}
\arguments{
\item{cp}{commercial paper data as retrieved by \code{getCommercialPaper()}}

\item{progressBar}{optional progress bar for \code{llply} progress monitoring,
default none; see {\code{plyr::create_progress_bar}}}
}
\value{
list containing data frame for each series, class \code{usfrbcpy}
}
\description{
Extract the commercial paper outstanding year end data set
}
\examples{
\dontrun{
cp <- getCommercialPaper()
yearend <- getCommercialPaperYearend(cp,"text")
summary(yearend)
}
}
\seealso{
\code{\link{getCommercialPaper}},
\code{\link{getCommercialPaperRates}},
\code{\link{getCommercialPaperRatesOld}},
\code{\link{getCommercialPaperVolumes}},
\code{\link{getCommercialPaperOutstanding}},
\code{\link{getCommercialPaperOutstandingOld}},
\code{\link{mergeSeries}}
}

