% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mergeSeries}
\alias{mergeSeries}
\title{Merge series}
\usage{
mergeSeries(data)
}
\arguments{
\item{data}{an extracted commercial paper series from rates, volume, outstanding, etc.}
}
\value{
if data has class \code{usfrbcpo}, which contains both monthly and
weekly series, then the return value is a list of two data frames
\code{weekly} and \code{monthly}; otherwise, the return value is a data frame.
}
\description{
Combines the series lists into a data frame with series as columns
and dates as row names.
}
\examples{
\dontrun{
cp <- getCommercialPaper()
rates <- getCommercialPaperRates(cp,"text")
table <- mergeSeries(rates)
head(table)
}
}

