% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{usfrbcp}
\alias{usfrbcp}
\alias{usfrbcp-package}
\title{Fetch US Federal Reserve Bank commercial paper history}
\description{
Fetches the US Federal Reserve Bank's commercial paper history and provides
transforms into data frames.
}
\details{
Downloads the US Federal Reserve Bank's commercial paper rate, volume, and
outstanding history in its original XML format, then translates that into
several data frames with data series as columns and dates as row names.
}
\author{
Matt Barry \email{mrb@softisms.com}
}
\references{
US Federal Reserve Bank resource web site
\url{http://www.federalreserve.gov/datadownload/}
}
\keyword{commercial}
\keyword{outstanding}
\keyword{paper}
\keyword{rate}
\keyword{volume}

