% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCommercialPaper}
\alias{getCommercialPaper}
\title{Get commercial paper}
\usage{
getCommercialPaper()
}
\value{
Returns the document represented as a list of class \code{usfrbcp}.
}
\description{
Fetch the US Federal Reserve Bank's Commercial Paper Data
}
\examples{
\dontrun{
cp <- getCommercialPaper()
summary(cp)
}
}
\author{
Matt Barry <mrb@softisms.com>
}

