% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCommercialPaperOutstandingOld}
\alias{getCommercialPaperOutstandingOld}
\title{Extract the commercial paper outstanding (old) data.}
\usage{
getCommercialPaperOutstandingOld(cp, progressBar = "none")
}
\arguments{
\item{cp}{commercial paper data as retrieved by \code{getCommercialPaper()}.}

\item{progressBar}{optional progress bar for \code{llply} progress monitoring,
default none; see {\code{plyr::create_progress_bar}}.}
}
\value{
Returns a list containing data frame for each series, class \code{usfrbcpw}.
}
\description{
Extracts the commercial paper outstanding (old) series from the commercial paper data set.
}
\examples{
\donttest{
cp <- getCommercialPaper()
outstanding <- getCommercialPaperOutstandingOld(cp,"text")
summary(outstanding)
}
}
\seealso{
\code{\link{getCommercialPaper}},
\code{\link{getCommercialPaperRates}},
\code{\link{getCommercialPaperRatesOld}},
\code{\link{getCommercialPaperVolumes}},
\code{\link{getCommercialPaperOutstanding}},
\code{\link{getCommercialPaperYearend}},
\code{\link{mergeSeries}}
}

