% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCommercialPaperRates}
\alias{getCommercialPaperRates}
\title{Extract commercial paper rates data.}
\usage{
getCommercialPaperRates(cp, progressBar = "none")
}
\arguments{
\item{cp}{commercial paper data as retrieved by \code{getCommercialPaper()}.}

\item{progressBar}{optional progress bar for \code{llply} progress monitoring,
default none; see {\code{plyr::create_progress_bar}}.}
}
\value{
Returns a list containing a data frame for each series, class \code{usfrbcpr}.
}
\description{
Extracts the commercial paper rates series from the commercial paper data set.
}
\examples{
\donttest{
cp <- getCommercialPaper()
rates <- getCommercialPaperRates(cp,"text")
summary(rates)
}
}
\seealso{
\code{\link{getCommercialPaper}},
\code{\link{getCommercialPaperRatesOld}},
\code{\link{getCommercialPaperVolumes}},
\code{\link{getCommercialPaperOutstanding}},
\code{\link{getCommercialPaperOutstandingOld}},
\code{\link{getCommercialPaperYearend}},
\code{\link{mergeSeries}}
}

