% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mergeSeries}
\alias{mergeSeries}
\title{Merge commercial paper series.}
\usage{
mergeSeries(data)
}
\arguments{
\item{data}{an extracted commercial paper series from rates, volume, outstanding, etc.}
}
\value{
If \code{data} has class \code{usfrbcpo}, which contains both monthly and
  weekly series, then the return value is a list of two data frames
  \code{weekly} and \code{monthly}; otherwise, the return value is a data frame.
}
\description{
Combines the commercial paper series lists into a data frame with series as columns
and dates as row names.
}
\examples{
\donttest{
cp <- getCommercialPaper()
rates <- getCommercialPaperRates(cp,"text")
rt <- mergeSeries(rates)
tail(rt)
}
}
\seealso{
\code{\link{getCommercialPaperRates}} for extracting commercial paper rates,
\code{\link{getCommercialPaperRatesOld}} for extracting old commercial paper rates,
\code{\link{getCommercialPaperVolumes}} for extracting commercial paper volumes,
\code{\link{getCommercialPaperOutstanding}} for extracting commercial paper outstanding,
\code{\link{getCommercialPaperOutstandingOld}} for extracting old commercial paper outstanding,
\code{\link{getCommercialPaperYearend}} for extracting commercial paper outstanding at year end.
}

